--- 01Knapsack/Input/knapsack_instances.csv
@@ -1,11 +1,0 @@
-10,10
-2,10
-1,5
-0,4
-0,2
-0,7
-0,2
-7,50
-0,2
-0,7
-0,5
